Добавляет фото пользователей ДЗ-7
