+++ js/main.js
@@ -1,3 +1,1 @@
-import {photosStock} from './data.js';
-
-console.log (photosStock);
+import './renderPhotos';
